requirements.txt
# ═══════════════════════════════════════════════════════
# UniversalBooks — Python Dependencies
# ═══════════════════════════════════════════════════════

# LiveKit Agent Framework + Plugins
livekit-agents[sarvam,anthropic,silero]>=1.4
livekit-api>=0.8.0

# Web Server (server.py)
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0

# HTTP Client (webhook calls in agent tools)
httpx>=0.27

# Environment variable management
python-dotenv>=1.0

# Groq plugin (backup LLM)
livekit-plugins-groq>=0.4
